Move Condition class to deposit-util, under o.d.deposit.util.async.
