+++ deposit-messaging/src/main/java/org/dataconservancy/pass/deposit/messaging/runner/ListenerRunner.java
@@ -19,7 +19,7 @@ import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
 import org.dataconservancy.pass.deposit.messaging.config.JmsConfig;
-import org.dataconservancy.pass.deposit.messaging.support.Condition;
+import org.dataconservancy.deposit.util.async.Condition;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.BeansException;
