+++ deposit-util/src/main/java/org/dataconservancy/deposit/util/async/Condition.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataconservancy.pass.deposit.messaging.support;
+package org.dataconservancy.deposit.util.async;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
